nrkcache 1.0
The script now accepts the URL of the NRK VOD page as input, making it relatively easy to use. There are options to specify the intended quality setting and to retrieve non-video content (useful for radio shows). If available, the subtitles will automatically be downloaded and converted to .srt for easy use with e.g. VLC (this feature is based on some anonymous Gist). Additionally, a text file is created with a short report of the video's meta data.