Update for new metdata format used by NRK
Rather than parsing the info we need right out of the HTML source code, we now use NRK's JSON metadata. Also, we retrieve the playlist of songs for radio broadcasts if available, and we create a .webloc file, which provides a quick way back to the NRK web page for the cached broadcast for OS X users.

There is also a new --base option in an effort to make this script work with non-NRK sources or partial caches, but this is somewhat 'hacky'. Use at your own risk.