Add option to create a new directory
So far, the script would use an existing directory, irrespective of whether or not it was empty. This could easily result in caches being created in the current working directory instead of whereever it was intended. This change offers a new --mkdir option, which creates a directory with the supplied name.

Additionally there is a new --comment option (which is always ignored) and minor improvements to verbosity.