Add option to cache a video's numbered part.
This is important for very long programmes like "Sommertoget minutt for minutt", which tend to be split into several parts by NRK at the two-hour mark. Earlier versions of this script would the only cache the first part, with the second part only being available after lengthy fiddling with the infamous --base option. This change offers an easy-to-use --part option to solve that problem.

Additionally, subtitles are marked as being in Norwegian language.