Hard-code media element API template
NRK moved the API template from the HTML out to some JavaScript code. But the template seems stable from the API documentation, so there is probably no downside in hard-coding it. The API base URL is also hard-coded as a fall-back (whether -we or -ne is used doesn't seem to matter).