Improve report formatting
The new NRK URLs no longer always contain the program ID. To better preserve this crucial piece of information, it is now appended to the description in the first lines of the report. The title is now shown in a line of its own before the rest of the report, which is now sorted alphabetically.