Allow separate video/audio streams (NRK fix)
NRK has recently started to only offer video and audio as technically separate streams. The format selection didn't account for that.

This solution tells the dlownloading tool to rewrite the output combining video and audio, then that output is still rewritten to combine it with subtitles as before. A better solution might be to first check if a combined video is available, and if not, download the parts separately, so we get to combine them ourselves and skip one rewriting step. But that'd require a bit of a redesign, which I don't wish to deal with right now. Hence this quick fix.

That said, a way to avoid that resdesign might be to rely on the assumption that for separately downloaded files, video will always be larger than audio. That would mean we could just use "," (instead of "+") for downloading both files, then parse the dir contents to figure out how to call ffmpeg. Hmm. Might be worth consideration.